--- coronahelpmap-master/CoronaHelp/firestore.indexes.json
@@ -1,4 +1,0 @@
-{
-  "indexes": [],
-  "fieldOverrides": []
-}
