--- coronahelpmap-master/CoronaHelp/functions/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
-  "scripts": {
-    "serve": "firebase serve --only functions",
-    "shell": "firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "dependencies": {
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.3.0",
-    "node-geocoder": "^3.25.1",
-    "node-open-geocoder": "^4.0.1",
-    "nodemailer": "^6.4.5"
-  },
-  "devDependencies": {
-    "firebase-functions-test": "^0.1.6"
-  },
-  "private": true
-}
