--- coronahelpmap-master/CoronaHelp/public/ReleaseNotes.md
@@ -1,24 +1,0 @@
-@1.3.0
-* **Added:** Can use a callback function for each marker infowindow close button element clicked.
-* **Changed:** Code Review.
-
------
-@1.2.0
-* **Changed:** Code Review.
-
------
-
-@1.1.0
-* **Added:** Can use a callback function for each marker element clicked.
-
------
-
-@1.0.0
-* Hello world! I am Markers On Map.
-
------
-
-@0.0.1-beta.01
-* I am trying to say "Hello World!"
-
------
