--- coronahelpmap-master/CoronaHelp/public/imprint.html
@@ -1,81 +1,0 @@
-﻿<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-
-    <title>Corona Helpmap - Imprint</title>
-    <meta name="description" content="Corona Helpmap" />
-
-    <meta name="content-language" content="en" />
-    <meta name="language" content="English" />
-    <meta name="robots" content="index, follow, noodp, noydir" />
-    <meta name="googlebot" content="index, follow, noodp, noydir" />
-
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
-    <link rel="shortcut icon" href="favicon.png" />
-
-    <link rel="canonical" href="https://furcan.github.io/Markers-On-Map/" />
-
-    <meta name="abstract" content="Corona Helpmap - Imprint">
-    <meta name="author" content="Various" />
-    <meta name="generator" content="Various" />
-    <meta name="designer" content="Various">
-    <meta name="copyright" content="© 2020 Corona Helpmap. All Rights Reserved.">
-
-    <meta name="theme-color" content="#DE2A00" />
-    
-    <link rel="dns-prefetch" href="//fonts.googleapis.com" />
-    <link id="FontMuli" href="https://fonts.googleapis.com/css?family=Muli:300,400,600,700,900&display=swap&subset=latin-ext" rel="stylesheet">
-    <link id="FontConsolas" href="https://fonts.googleapis.com/css?family=Inconsolata:400,700&amp;subset=latin-ext" rel="stylesheet">
-
-    <link rel="stylesheet" href="imprint.css?v=1.3.0">
-
-</head>
-
-<body>
-
-    <div id="Wrapper">
-
-        <footer id="Footer">
-            <p><b><span style="font-size: small;">Impressum</span></b><br /><br /><span>Angaben gem&auml;&szlig; &sect; 5 TMG</span><br /><br /><span>XYZ XYZ</span><br /><span>XYZ</span><br /><span>XYZ XYZ</span><br /><br /><span>Kontakt</span><br /><span>Telefon: xyz</span><br /><span>E-Mail:&nbsp;</span><a href="mailto:xyz" target="_blank" rel="noopener">XYZ</a><span></span><br /><br /><span>EU-Streitschlichtung</span><br /><span>Die Europ&auml;ische Kommission stellt eine Plattform zur Online-Streitbeilegung (OS) bereit:&nbsp;</span><a href="https://ec.europa.eu/consumers/odr" target="_blank" data-saferedirecturl="https://www.google.com/url?q=https://ec.europa.eu/consumers/odr&amp;source=gmail&amp;ust=1584743915410000&amp;usg=AFQjCNFWQGKBQCX4XRfN9idzb-BN3mZJFw" rel="noopener">https://ec.europa.eu/<wbr />consumers/odr</a><span>.</span><br /><span>Unsere E-Mail-Adresse finden Sie oben im Impressum.</span><br /><br /><span>Verbraucherstreitbeilegung/</span><wbr /><span>Universalschlichtungsstelle</span><br /><span>Wir sind nicht bereit oder verpflichtet, an Streitbeilegungsverfahren vor einer Verbraucherschlichtungsstelle teilzunehmen.</span><br /><br /><span>Haftung f&uuml;r Inhalte</span><br /><span>Als Diensteanbieter sind wir gem&auml;&szlig; &sect; 7 Abs.1 TMG f&uuml;r eigene Inhalte auf diesen Seiten nach den allgemeinen Gesetzen verantwortlich. Nach &sect;&sect; 8 bis 10 TMG sind wir als Diensteanbieter jedoch nicht verpflichtet, &uuml;bermittelte oder gespeicherte fremde Informationen zu &uuml;berwachen oder nach Umst&auml;nden zu forschen, die auf eine rechtswidrige T&auml;tigkeit hinweisen.</span><br /><span>Verpflichtungen zur Entfernung oder Sperrung der Nutzung von Informationen nach den allgemeinen Gesetzen bleiben hiervon unber&uuml;hrt. Eine diesbez&uuml;gliche Haftung ist jedoch erst ab dem Zeitpunkt der Kenntnis einer konkreten Rechtsverletzung m&ouml;glich. Bei Bekanntwerden von entsprechenden Rechtsverletzungen werden wir diese Inhalte umgehend entfernen.</span></p>
-        </footer>
-
-        
-
-            
-        </div>
-
-        
-
-
-
-        <footer id="Footer">
-            <p class="copyright">&copy; 2020 Corona Helpmap. All rights reserved.</p>
-            <a class="github" target="_blank" rel="noopener" href="https://coronahelpmap.com/">Imprint</a>
-        </footer>
-
-    </div>
-    <script src="https://www.gstatic.com/firebasejs/7.11.0/firebase-app.js"></script>
-    <script src="/__/firebase/7.11.0/firebase-firestore.js"></script>
-
-
-    <script src="https://cdn.jsdelivr.net/npm/notiflix/dist/notiflix-aio-1.8.0.min.js"></script>
-    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-    <script src="https://kit.fontawesome.com/833245c2e4.js"></script>
-    <script src="index-tooltip.js"></script>
-    <script src="dist/markers-on-map-1.3.0.js"></script>
-    <script src="index.js?v=1.3.0"></script>
-    
-    <!-- The core Firebase JS SDK is always required and must be listed first -->
-
-<!-- TODO: Add SDKs for Firebase products that you want to use
-     https://firebase.google.com/docs/web/setup#available-libraries -->
-
-
-
-
-
-</body>
-
-</html>
