--- coronahelpmap-master/CoronaHelp/public/send.html
@@ -1,138 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="utf-8">
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link href="https://fonts.googleapis.com/css?family=Roboto:400,700" rel="stylesheet">
-<link id="FontMuli" href="https://fonts.googleapis.com/css?family=Muli:300,400,600,700,900&display=swap&subset=latin-ext" rel="stylesheet">
-
-<title>Nachricht senden</title>
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script> 
-<style type="text/css">
-    body{
-        color: #fff;
-        background: #DE2A00;
-        font-family: 'Muli', sans-serif;
-    }
-    .form-control{
-        height: 41px;
-        background: #f2f2f2;
-        box-shadow: none !important;
-        border: none;
-    }
-    .form-control:focus{
-        background: #e2e2e2;
-    }
-    .form-control, .btn{        
-        border-radius: 3px;
-    }
-    .signup-form{
-        width: 390px;
-        margin: 30px auto;
-    }
-    .signup-form form{
-        color: #999;
-        border-radius: 3px;
-        margin-bottom: 15px;
-        background: #fff;
-        box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-        padding: 30px;
-    }
-    .signup-form h2 {
-        color: #333;
-        font-weight: bold;
-        margin-top: 0;
-    }
-    .signup-form hr {
-        margin: 0 -30px 20px;
-    }    
-    .signup-form .form-group{
-        margin-bottom: 20px;
-    }
-    .signup-form input[type="checkbox"]{
-        margin-top: 3px;
-    }
-    .signup-form .row div:first-child{
-        padding-right: 10px;
-    }
-    .signup-form .row div:last-child{
-        padding-left: 10px;
-    }
-    .signup-form .btn{        
-        font-size: 16px;
-        font-weight: bold;
-        background: #3598dc;
-        border: none;
-        min-width: 140px;
-    }
-    .signup-form .btn:hover, .signup-form .btn:focus{
-        background: #2389cd !important;
-        outline: none;
-    }
-    .signup-form a{
-        color: #fff;
-        text-decoration: underline;
-    }
-    .signup-form a:hover{
-        text-decoration: none;
-    }
-    .signup-form form a{
-        color: #3598dc;
-        text-decoration: none;
-    }   
-    .signup-form form a:hover{
-        text-decoration: underline;
-    }
-    .signup-form .hint-text {
-        padding-bottom: 15px;
-        text-align: center;
-    }
-</style>
-</head>
-<body>
-
-<div class="signup-form">
-    <form id="input_fields">
-        <h2>Nachricht senden</h2>
-        <p>Du bekommst umgehend Bescheid.</p>
-        <hr>
-        
-        <div class="form-group">
-            <div class="row">
-                <div class="col-xs-6"><input type="text" class="form-control" name="first_name" placeholder="Vorname" required="required"></div>
-                <div class="col-xs-6"><input type="text" class="form-control" name="last_name" placeholder="Nachname" required="required"></div>
-            </div>          
-        </div>
-        <div class="form-group">
-            <input type="text" class="form-control" name="address" placeholder="Postleitzahl" required="required">
-        </div>
-        <div class="form-group">
-            <input type="email" class="form-control" name="contactInfo" placeholder="Email" required="required">
-        </div> 
-        <div class="form-group">
-            <textarea type="text" class="form-control" name="typeOfHelp" placeholder="Nachricht" required="required" rows="6"></textarea>
-        </div>
-        
-    
-        <div class="form-group">
-            <label class="checkbox-inline"><input type="checkbox" required="required"> Ich akzeptiere hiermit die <a href="#">Nutzungs- </a> &amp; <a href="#">Datenschutz-Bestimmungen
-</a></label>
-        </div>
-        
-        <div class="form-group">
-            <button type="submit" class="btn btn-primary btn-lg">Nachricht senden</button>
-        </div>
-    </form>
-    <div class="hint-text">Angebot wieder von der Karte entfernen? <a href="#">Hier wieder abmelden</a></div>
-</div>
-        <script src="https://www.gstatic.com/firebasejs/7.12.0/firebase-app.js"></script>
-        <script src="https://www.gstatic.com/firebasejs/7.12.0/firebase-firestore.js"></script>
-        <script src="/__/firebase/init.js"></script>
-        <script src="send.js"></script>
-</body>
-</html>                            
-
-
