--- coronahelpmap-master/README.md
@@ -1,13 +1,0 @@
-## Über Corona Helpmap
-
-Die [Corona Helpmap](https://www.coronahelpmap.com) dient zur direkten Nachbarschaftshilfe.
-
-## Über #wirvsvirus
-
-Dieses Projekt würde als Beitrag zum [#wirvsvirus Hackathon](https://wirvsvirushackathon.org/) der deutschen Bundesregierung entwickelt.
-
-## Probier's aus
-
-Besuche [Corona Helpmap](https://www.coronahelpmap.com) um den Prototypen zu testen.
-
-
